add exception handler

+++ src/main/java/cn/lxh/community/controller/QuestionController.java
@@ -10,12 +10,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 @Controller
 public class QuestionController {
-
     @Autowired
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id,
+    public String question(@PathVariable(name = "id") Long id,
                            Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         questionService.incView(id);
